add documentation router
